+++ src/Enums.js
@@ -1,0 +1,14 @@
+
+// Exports a bunch of symbols^H^H^H^H^H ints for use as enums
+
+export var INTERVAL = 1;
+export var MINIMAL = 2;
+export var INSTANT = 3;
+export var FRAME = 4;
+export var FASTFRAME = 5;
+
+// export var INTERVAL = new Symbol();
+// export var MINIMAL = new Symbol();
+// export var INSTANT = new Symbol();
+//
+